Bert_Patent_Classification
Data preprocessing
Data preprocessing involved several steps crucial in the preparation of textual data before inputting it into the BERT uncased model. Inherent in this process was tokenization. Using the BERT tokenizer from Hugging Face, we turned the abstracts of patents into tokens readable to the model. This step resized long sequences and padded the short ones to procure a uniform input size, which is always required by the BERT model. In doing so, we maintained the integrity of the text data of interest while making it compatible with the architecture of the model.
It implemented dynamic padding with the help of the data collator because all patent abstracts were not of equal length. This would ensure that each input batch to be padded to the length of its longest sequence in the batch, optimizing memory usage during training and computational efficiency. This step was particularly important to handle the diverse nature of the patent abstracts effectively.
Bert Model training
The training of the 'BERT Uncased model' for fine-tuning also necessitated proper design and implementation for its adaptability to patent classification. The fine-tuning parameters had to be configured using the Hugging Face model, setting some specific training arguments on optimization, in a manner allowing maximum learning and robust evaluation. 
The TrainingArguments class was used in setting crucial parameters. The output directory (OUT_DIR) where the checkpoints of trained models will be saved was set. One selected learning rate of 5e-5 to hit a superposition for learning speed and stability. The same batch size, 32, was used for both training and evaluation to allow efficient processing and keep memory constraints at bay. The target number of maximal training epochs was set at 7, allowing the model every opportunity to learn the fine-grained patterns present within the patent texts. The regularization I applied to the model was weight decay of 0.01 to avoid overfitting. Performance evaluation takes place at the end of each epoch of the model. At this point, model states are saved such that the different stages during training can be compared. At the end of this training process, the best-performing model from this training was then loaded; therefore, the final model from this training process represented the optimal state reached during the trainer. Also, the training process was saved in history to further be visualized, which means one could help find an idea of how the model learns to do his task. Training in half-precision mode is on to utilize the computational capabilities of the modern hardware. This technique speeds up training, thanks to lower memory utilization and faster computations, yet it does not compromise the accuracy of the model. In this example, I used the Trainer class from Hugging Face, which made training very easy because it just required the model and training arguments tobe given along with the datasets for training and evaluation. A custom metric for accuracy was loaded using the evaluate library. A function for computing metrics was defined in order to relate the model's predictions with the true labels, ensuing a clear measure of classification performance. As shown in the training and validation loss and accuracy graph, the training loss was seen to decrease consistently, thereby proving effective learning. In the validation loss, a slight increment was seen after initial epochs, which is a sign of overfitting. The accuracy was remaining quite stable and reached the level of 65.18% in the second epoch.